Move helpers and selectors into same subfolder as clusters integration tests
