+++ ui/apps/platform/cypress/constants/apiEndpoints.js
@@ -50,11 +50,7 @@ export const alerts = {
 };
 
 export const clusters = {
-    single: 'v1/clusters/**',
     list: 'v1/clusters',
-    clusterDefaults: '/v1/cluster-defaults',
-    sensorUpgradesConfig: '/v1/sensorupgrades/config',
-    zip: 'api/extensions/clusters/zip',
 };
 
 export const risks = {
