+++ ui/apps/platform/cypress/integration/systemHealth/clusters.test.js
@@ -1,7 +1,6 @@
-import { clustersUrl } from '../../constants/ClustersPage';
 import { selectors } from '../../constants/SystemHealth';
 import withAuth from '../../helpers/basicAuth';
-import { interactAndVisitClusters } from '../../helpers/clusters';
+import { interactAndVisitClusters } from '../clusters/Clusters.helpers';
 import { setClock, visitSystemHealth } from '../../helpers/systemHealth';
 
 function visitSystemHealthWithClustersFixtureFilteredByNames(fixturePath, clusterNames) {
@@ -23,7 +22,6 @@ describe('System Health Clusters without fixture', () => {
         interactAndVisitClusters(() => {
             cy.get(selectors.clusters.viewAllButton).click();
         });
-        cy.location('pathname').should('eq', clustersUrl);
     });
 });
